Add single product page
Add single product page based on slugs. This is a work in progress, as
it's very expensive to build a "pretty" url without making two API
calls--one for the static path to create the URLs, and another within
the `getStaticProps` to find the product by running a full product API
call and comparing each product's slugified name with the slug
parameter.

An optimal approach would be to either:

  1. Don't use pretty URLs, and instead use the product ID, which can be
     queried against in the Shopware API
  2. Create a new module to add a slug to the response object of the
     `/product` API route